CART-523 array: use consistent variable type
The ca_count member is defined as size_t.

struct crt_array {
	size_t		 ca_count;
	void		*ca_arrays;
};

But inside marshaling function its used as uint64_t.

hg_ret = hg_proc_hg_uint64_t(proc, &array->ca_count);

So, there is assumption that size_t has the same size as uint64_t.
To avoid an issue on machines where this is not true we should
change definition to "uint64_t ca_count".

Reviewed-on: https://review.hpdd.intel.com/33338
Tested-by: Jenkins

+++ src/include/cart/types.h
@@ -210,7 +210,7 @@ struct crt_req_format {
 };
 
 struct crt_array {
-	size_t		 ca_count;
+	uint64_t	 ca_count;
 	void		*ca_arrays;
 };
